ai/pyproject.toml
[project]
name = "ai"
version = "1.0.0"
authors = [
    {name = "Samay Nandan",email = "samaynandan@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
ruff = "^0.11.11"
mypy = "^1.15.0"
pyinstaller = "^6.13.0"
python-dotenv = "^1.1.0"

[tool.mypy]
explicit_package_bases = true
python_version = "3.11"
warn_unused_configs = true
files = ["src"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.ruff]
line-length = 88
target-version = "py311"
src = ["src"]
